Refactored all to java package com.slytechs.jnet.jnetruntime.systable
