+++ pom.xml
@@ -1,166 +1,19 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<groupId>com.slytechs.jnet</groupId>
-	<artifactId>system-tables</artifactId>
-	<version>1.0-SNAPSHOT</version>
-	<name>System Tables</name>
-	<description>Access to various system tables in Java</description>
-	<url>https://www.slytechs.com</url>
-	<properties>
-		<jdk.version>19</jdk.version>
-		<project.build.sourceEncoding>${java.encoding}</project.build.sourceEncoding>
-		<maven.compiler.source>${jdk.version}</maven.compiler.source>
-		<maven.compiler.target>${jdk.version}</maven.compiler.target>
-		<java.preview>--enable-preview</java.preview>
-		<java.encoding>UTF-8</java.encoding>
-	</properties>
-	<licenses>
-		<license>
-			<name>Free License</name>
-			<url>https://www.slytechs.com/free-license</url>
-		</license>
-	</licenses>
-	<organization>
-		<name>Sly Technologies Inc</name>
-		<url>http://slytechs.com</url>
-	</organization>
-	<developers>
-		<developer>
-			<id>markb</id>
-			<name>Mark Bednarczyk</name>
-			<email>mark@slytechs.com</email>
-			<organization>Sly Technlogies</organization>
-			<organizationUrl>http://slytechs.com</organizationUrl>
-		</developer>
-	</developers>
-	<dependencies>
-		<!-- https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-compiler-plugin -->
-		<dependency>
-			<groupId>org.apache.maven.plugins</groupId>
-			<artifactId>maven-compiler-plugin</artifactId>
-			<version>3.10.1</version>
-		</dependency>
-		<dependency>
-			<groupId>org.junit.jupiter</groupId>
-			<artifactId>junit-jupiter-engine</artifactId>
-			<version>5.9.1</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.slytechs.jnet</groupId>
-			<artifactId>jnet-runtime</artifactId>
-			<version>1.0.0-SNAPSHOT</version>
-		</dependency>
-	</dependencies>
-	<build>
-		<resources>
-			<resource>
-				<directory>src/main/java</directory>
-				<excludes>
-					<exclude>**/*.java</exclude>
-				</excludes>
-			</resource>
-		</resources>
-		<plugins>
-			<plugin>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.10.1</version>
-				<configuration>
-					<source>${jdk.version}</source>
-					<target>${jdk.version}</target>
-					<release>${jdk.version}</release>
-					<enablePreview>true</enablePreview>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-site-plugin</artifactId>
-				<version>3.7.1</version>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-javadoc-plugin</artifactId>
-				<version>3.4.1</version>
-				<configuration>
-					<doctitle>Public API for ${project.name} ${project.version}</doctitle>
-					<windowtitle>Public API for ${project.name} ${project.version}</windowtitle>
-					<encoding>${java.encoding}</encoding>
-					<source>${jdk.version}</source>
-					<additionalOptions>
-								${java.preview}
-					</additionalOptions>
-					<show>public</show>
-					<failOnError>false</failOnError>
-				</configuration>
-			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-scm-publish-plugin</artifactId>
-				<version>3.0.0</version>
-				<configuration>
-					<scmBranch>gh-pages</scmBranch>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-	<reporting>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-javadoc-plugin</artifactId>
-				<version>3.4.1</version>
-				<configuration>
-					<doctitle>Public API for ${project.name} ${project.version}</doctitle>
-					<windowtitle>Public API for ${project.name} ${project.version}</windowtitle>
-					<testDoctitle>Test API for ${project.name} ${project.version}</testDoctitle>
-					<testWindowtitle>Test API for ${project.name} ${project.version}</testWindowtitle>
-					<encoding>${java.encoding}</encoding>
-					<source>${jdk.version}</source>
-					<additionalOptions>
-								${java.preview}
-					</additionalOptions>
-				</configuration>
-				<reportSets>
-					<reportSet>
-						<id>aggregate-no-fork</id>
-						<configuration>
-							<goal>aggregate-no-fork</goal>
-							<show>public</show>
-							<skip>false</skip>
-						</configuration>
-						<reports>
-							<report>javadoc</report>
-						</reports>
-					</reportSet>
-					<reportSet>
-						<id>test-aggregate-no-fork</id>
-						<configuration>
-							<goal>test-aggregate-no-fork</goal>
-							<show>private</show>
-							<skip>true</skip>
-						</configuration>
-						<reports>
-							<report>test-javadoc</report>
-						</reports>
-					</reportSet>
-				</reportSets>
-			</plugin>
-		</plugins>
-	</reporting>
-	<issueManagement>
-		<system>Github Issue Tracking</system>
-		<url>https://github.com/slytechs-repos/system-tables/issues</url>
-	</issueManagement>
-	<scm>
-		<connection>scm:git:git://github.com/slytechs-repos/system-tables.git</connection>
-		<developerConnection>scm:git:ssh:git://github.com/slytechs-repos/system-tables.git</developerConnection>
-		<url>https://github.com/slytechs-repos/system-tables</url>
-	</scm>
-	<distributionManagement>
-		<!-- mvn clean site site:stage scm-publish:publish-scm -->
-		<site>
-			<id>github</id>
-			<url>scm:git:git@github.com:slytechs-repos/system-tables.git</url>
-		</site>
-	</distributionManagement>
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.slytechs.jnet.jnetruntime</groupId>
+    <artifactId>jnetruntime-sdk-parent</artifactId>
+    <version>0.0.1-SNAPSHOT</version>
+  </parent>
+  <artifactId>jnetruntime-systable</artifactId>
+  <name>jNetRuntime Systable</name>
+  <description>Access to system tables</description>
+  
+  <dependencies>
+	  <dependency>
+		  <groupId>com.slytechs.jnet.jnetruntime</groupId>
+		  <artifactId>jnetruntime-api</artifactId>
+		  <version>${jnetruntime-sdk.version}</version>
+	  </dependency>
+  </dependencies>
 </project>
