+++ src/main/java/com/slytechs/jnet/jnetruntime/systable/dns/DnsRecord.java
@@ -15,7 +15,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.slytechs.jnet.systable.dns;
+package com.slytechs.jnet.jnetruntime.systable.dns;
 
 /**
  * @author Sly Technologies Inc
