+++ src/main/java/com/slytechs/jnet/jnetruntime/systable/dns/DnsTable.java
@@ -15,13 +15,13 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.slytechs.jnet.systable.dns;
+package com.slytechs.jnet.jnetruntime.systable.dns;
 
 import java.nio.file.Path;
 import java.util.Optional;
 
-import com.slytechs.jnet.systable.SystemTable;
-import com.slytechs.protocol.runtime.NotFound;
+import com.slytechs.jnet.jnetruntime.NotFound;
+import com.slytechs.jnet.jnetruntime.systable.SystemTable;
 
 /**
  * @author Sly Technologies Inc
@@ -39,7 +39,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#name()
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#name()
 	 */
 	@Override
 	public Optional<String> name() {
@@ -47,7 +47,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#description()
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#description()
 	 */
 	@Override
 	public Optional<String> description() {
@@ -55,7 +55,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#path()
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#path()
 	 */
 	@Override
 	public Optional<Path> path() {
@@ -63,7 +63,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#contains(java.lang.Object)
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#contains(java.lang.Object)
 	 */
 	@Override
 	public boolean contains(DnsRecord t) {
@@ -71,7 +71,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#find()
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#find()
 	 */
 	@Override
 	public Optional<DnsRecord> find() {
@@ -79,7 +79,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#get()
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#get()
 	 */
 	@Override
 	public DnsRecord get() throws NotFound {
@@ -87,7 +87,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#add(java.lang.Object)
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#add(java.lang.Object)
 	 */
 	@Override
 	public boolean add(DnsRecord t) {
@@ -95,7 +95,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#set(java.lang.Object)
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#set(java.lang.Object)
 	 */
 	@Override
 	public boolean set(DnsRecord t) {
@@ -103,7 +103,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#remove(java.lang.Object)
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#remove(java.lang.Object)
 	 */
 	@Override
 	public boolean remove(DnsRecord t) throws NotFound {
@@ -111,7 +111,7 @@ public class DnsTable implements SystemTable<DnsRecord> {
 	}
 
 	/**
-	 * @see com.slytechs.jnet.systable.SystemTable#disable(java.lang.Object)
+	 * @see com.slytechs.jnet.jnetruntime.systable.SystemTable#disable(java.lang.Object)
 	 */
 	@Override
 	public boolean disable(DnsRecord t) throws NotFound {
